% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{distance}
\alias{distance}
\title{Measure the distance between an \code{R} plot and a plot saved as a PNG file.}
\usage{
distance(x, path, tolerance = 0, metric = "AE", fuzz = 0, ...)
}
\arguments{
\item{x}{a \code{ggplot2} object or a function which returns a base \code{R} plot when executed.}

\item{path}{Path to the saved plot in PNG format.}

\item{tolerance}{distance estimates larger than this threshold will trigger a test failure. Scale depends on the \code{metric} argument. With the default \code{metric="AE"} (absolute error), the \code{tolerance} corresponds roughly to the number of pixels of difference between the plot and the reference image. See also \code{?distance_plot}}

\item{metric}{string with a \href{http://www.imagemagick.org/script/command-line-options.php#metric}{metric}
from \link[magick:metric_types]{metric_types()} such as \code{"AE"} or \code{"phash"}}

\item{fuzz}{relative color distance (value between 0 and 100) to be considered similar
in the filling algorithm}
}
\description{
Measure the distance between an \code{R} plot and a plot saved as a PNG file.
}
