% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_pdiff.R
\name{expect_pdiff}
\alias{expect_pdiff}
\title{Test if printed output matches a target printout}
\usage{
expect_pdiff(x, label, disp.width = 200)
}
\arguments{
\item{x}{an object which returns text to the console when calling \verb{print(x})`}

\item{label}{a string to identify the test (alpha-numeric, hypohens, or underscores). Each plot in the test suite must have a unique label.}

\item{disp.width}{integer(1L) number of display columns to take up; note that
in \dQuote{sidebyside} \code{mode} the effective display width is half this
number (set to 0L to use default widths which are \code{getOption("width")}
for normal styles and \code{80L} for HTML styles.  Future versions of
\code{diffobj} may change this to larger values for two dimensional objects
for better diffs (see details).}
}
\value{
A \code{tinytest} object. A tinytest object is a
\code{logical} with attributes holding information about the test that was
run
}
\description{
This expectation can be used with \code{tinytest} to check if the new plot matches
a target plot.
}
\details{
When the expectation is checked for the first time, the expectation fails and
a reference text file is saved to the \verb{inst/tinytest/_tinyviztest} folder.

To update a snapshot, delete the reference file from the \verb{_tinyviztest}
folder and run the test suite again.
}
