% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyvizreview.R
\name{tinyvizreview}
\alias{tinyvizreview}
\title{Review the plots rendered by the \code{tinyviztest} suite}
\usage{
tinyvizreview(
  label = NULL,
  path = "inst/tinytest/_tinyviztest_review",
  width = 12,
  height = width/3
)
}
\arguments{
\item{label}{a string to identify the test. Each plot in the test suite must have a unique label. When \code{label} is \code{NULL}, the function returns a list of images ready for review.}

\item{path}{Path to the \verb{_tinyviztest_review} folder. This folder is created when expectations fail while using \code{tinytest}.}

\item{width}{Width of the graphics device window}

\item{height}{Height of the graphics device window}
}
\value{
A plot with 3 panels: old, new, and diff pictures.
}
\description{
Review the plots rendered by the \code{tinyviztest} suite
}
