\name{tinysnapshot-package}
\alias{tinysnapshot-package}
\alias{tinysnapshot}
\docType{package}
\title{\packageTitle{tinysnapshot}}
\description{\packageDescription{tinysnapshot}}
\section{Package Content}{\packageIndices{tinysnapshot}}
\author{\packageAuthor{tinysnapshot}}
\section{Maintainer}{\packageMaintainer{tinysnapshot}}
\keyword{package}
