\name{tinyviztest-package}
\alias{tinyviztest-package}
\alias{tinyviztest}
\docType{package}
\title{\packageTitle{tinyviztest}}
\description{\packageDescription{tinyviztest}}
\section{Package Content}{\packageIndices{tinyviztest}}
\author{\packageAuthor{tinyviztest}}
\section{Maintainer}{\packageMaintainer{tinyviztest}}
\keyword{package}
